# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "The core API for DocPad, managing patient records, scheduling, and prescriptions for the Bangladeshi demographic.",
        "title": "DocPad Hospital Management API",
        "contact": {
            "name": "DocPad Support",
            "email": "support@docpad.bd"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Authenticate user with email and password to receive a JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "User Login",
                "parameters": [
                    {
                        "description": "Login Credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request or invalid data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Blacklists the provided refresh token to end the session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout User",
                "parameters": [
                    {
                        "description": "Refresh Token to revoke",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LogoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logout success",
                        "schema": {
                            "$ref": "#/definitions/jsonutil.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/jsonutil.Response"
                        }
                    }
                }
            }
        },
        "/auth/rotate": {
            "post": {
                "description": "Generates a new access and refresh token pair using a valid refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Rotate Tokens",
                "parameters": [
                    {
                        "description": "Refresh Token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RotateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rotation success",
                        "schema": {
                            "$ref": "#/definitions/domain.Tokenpair"
                        }
                    },
                    "401": {
                        "description": "Token revoked or invalid",
                        "schema": {
                            "$ref": "#/definitions/jsonutil.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Tokenpair": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "refreshToke": {
                    "type": "string"
                }
            }
        },
        "dto.AuthRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "hehe@gmail.com"
                },
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "hehe1234"
                }
            }
        },
        "dto.LogoutRequest": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJFUzI1NiIsInR5c..."
                }
            }
        },
...
        "dto.RotateRequest": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJFUzI1NiIsInR5c..."
                }
            }
        },
...
        "jsonutil.ErrorItem": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "e.g., \"INVALID_EMAIL\"",
                    "type": "string"
                },
                "field": {
                    "description": "e.g., \"email\"",
                    "type": "string"
                },
                "message": {
                    "description": "User-friendly description",
                    "type": "string"
                }
            }
        },
        "jsonutil.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Primary response data"
                },
                "errors": {
                    "description": "Error details",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/jsonutil.ErrorItem"
                    }
                },
                "message": {
                    "description": "Human-readable message",
                    "type": "string"
                },
                "meta": {
                    "description": "Pagination, etc."
                },
                "status": {
                    "description": "\"success\", \"fail\", or \"error\"",
                    "type": "string"
                },
                "statusCode": {
                    "description": "HTTP status code",
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}